pyproject.toml
[project]
name = "mystroke-backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "B_K", email = "chayanongg2550@gmail.com" }
]
dependencies = [
    "fastapi>=0.103.2",
    "tensorflow>=2.13.1",
    "uvicorn[standard]>=0.23.2",
    "python-multipart>=0.0.6",
    "Pillow>=10.0.1",
    "mediapipe>=0.10.7",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
